backend/requirements.txt
# Core LiveKit and voice agent dependencies
livekit-agents[openai,silero,elevenlabs]==1.0.11
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
httpx==0.28.1

# Additional dependencies for our project
websockets==12.0
aiofiles==23.2.1
pydantic==2.5.3
python-multipart==0.0.9
PyJWT==2.8.0

# Development dependencies
ipython==8.13.2
pytest==7.4.4
pytest-asyncio==0.23.3